requirements.txt
opencv-python
Pillow
tk